+++ src/components/Button.vue
@@ -1,6 +1,6 @@
 <template>
   <button
-    class="border border-black px-2 py-1 text-xs font-medium shadow hover:bg-neutral-100 dark:border-white dark:hover:bg-neutral-700"
+    class="border border-black px-2 py-1 text-xs font-medium shadow hover:bg-neutral-100 disabled:cursor-not-allowed disabled:bg-neutral-100 disabled:text-neutral-500 disabled:shadow-none dark:border-white dark:hover:bg-neutral-700 dark:disabled:bg-neutral-700"
   >
     <slot></slot>
   </button>