+++ src/sections/__tests__/Licence.spec.ts
@@ -2,11 +2,10 @@ import { describe, it, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
 import Clipboard from 'v-clipboard'
 
-import type { AccessRestriction, Licence as LicenceT } from '@/types/app'
+import type { AccessRestriction } from '@/types/app'
 import type { Constraint } from '@/types/iso'
+import { getLicence } from '@/utils/data'
 import Licence from '@/sections/Licence.vue'
-
-import licencesData from '@/data/licences.json'
 
 const openAccessRestriction: AccessRestriction = {
   slug: 'anonymous',
@@ -23,8 +22,8 @@ const closedAccessRestriction: AccessRestriction = {
 }
 
 describe('Licence', () => {
-  it('renders licence from open choice', async () => {
-    const expectedLicence: LicenceT = licencesData['licences']['OGL_UK_3_0']
+  it('renders and emits licence from open choice', async () => {
+    const expectedLicence = getLicence('OGL_UK_3_0')
     const expectedConstraint: Constraint = {
       type: 'usage',
       restriction_code: 'license',
@@ -46,10 +45,18 @@ describe('Licence', () => {
     // don't need to wait for filter as initial licence happens to be open
 
     expect(wrapper.find('pre').text()).toBe(JSON.stringify(expectedConstraint, null, 2))
+
+    await wrapper.vm.$nextTick()
+
+    const emittedLicence: unknown[][] | undefined = wrapper.emitted('update:licence')
+    expect(emittedLicence).toBeTruthy()
+    if (emittedLicence) {
+      expect(emittedLicence[0][0]).toEqual(expectedLicence)
+    }
   })
 
-  it('renders licence from closed choice', async () => {
-    const expectedLicence: LicenceT = licencesData['licences']['X_FAKE_CLOSED']
+  it('renders and emits licence from closed choice', async () => {
+    const expectedLicence = getLicence('X_FAKE_CLOSED')
     const expectedConstraint: Constraint = {
       type: 'usage',
       restriction_code: 'license',
@@ -72,10 +79,16 @@ describe('Licence', () => {
     await wrapper.vm.$nextTick()
 
     expect(wrapper.find('pre').text()).toBe(JSON.stringify(expectedConstraint, null, 2))
+
+    const emittedLicence: unknown[][] | undefined = wrapper.emitted('update:licence')
+    expect(emittedLicence).toBeTruthy()
+    if (emittedLicence) {
+      expect(emittedLicence[0][0]).toEqual(expectedLicence)
+    }
   })
 
   it('updates licence from choice', async () => {
-    const expectedLicenceName = 'Creative Commons Attribution 4.0 International'
+    const expectedLicence = getLicence('CC_BY_4_0')
 
     const wrapper = mount(Licence, {
       props: {
@@ -91,13 +104,19 @@ describe('Licence', () => {
     // set radio input with id 'licence-CC_BY_4_0' to checked
     await wrapper.find('input#licence-CC_BY_4_0').setValue()
 
-    expect(wrapper.find('pre').text()).toContain(expectedLicenceName)
+    expect(wrapper.find('pre').text()).toContain(expectedLicence.name)
+
+    // check second emit for licence update
+    const emittedLicence: unknown[][] | undefined = wrapper.emitted('update:licence')
+    expect(emittedLicence).toBeTruthy()
+    if (emittedLicence) {
+      expect(emittedLicence[1][0]).toEqual(expectedLicence)
+    }
   })
 
   it('updates licence from restriction', async () => {
-    const expectedInitialLicenceHref =
-      'http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/'
-    const expectedUpdatedLicenceHref = 'http://www.example.com'
+    const expectedInitialLicence = getLicence('OGL_UK_3_0')
+    const expectedUpdatedLicence = getLicence('X_FAKE_CLOSED')
 
     const wrapper = mount(Licence, {
       props: {
@@ -110,11 +129,20 @@ describe('Licence', () => {
       },
     })
 
-    expect(wrapper.find('pre').text()).toContain(expectedInitialLicenceHref)
+    expect(wrapper.find('pre').text()).toContain(expectedInitialLicence.url)
+
+    const emittedLicence: unknown[][] | undefined = wrapper.emitted('update:licence')
+    expect(emittedLicence).toBeTruthy()
+    if (emittedLicence) {
+      expect(emittedLicence[0][0]).toEqual(expectedInitialLicence)
+    }
 
     // update prop to closed access restriction
     await wrapper.setProps({ accessRestriction: closedAccessRestriction })
 
-    expect(wrapper.find('pre').text()).toContain(expectedUpdatedLicenceHref)
+    expect(wrapper.find('pre').text()).toContain(expectedUpdatedLicence.url)
+    if (emittedLicence) {
+      expect(emittedLicence[1][0]).toEqual(expectedUpdatedLicence)
+    }
   })
 })